requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.2
cffi==2.0.0
click==8.3.1
fastapi==0.135.0
h11==0.16.0
idna==3.11
passlib==1.7.4
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-multipart==0.0.22
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
